Removed extra tsdb/testutil after merge.
