+++ tsdb/chunks/chunks_test.go
@@ -16,7 +16,7 @@ package chunks
 import (
 	"testing"
 
-	"github.com/prometheus/prometheus/tsdb/testutil"
+	"github.com/prometheus/prometheus/util/testutil"
 )
 
 func TestReaderWithInvalidBuffer(t *testing.T) {
