+++ tsdb/wal/wal_test.go
@@ -24,7 +24,7 @@ import (
 	"testing"
 
 	client_testutil "github.com/prometheus/client_golang/prometheus/testutil"
-	"github.com/prometheus/prometheus/tsdb/testutil"
+	"github.com/prometheus/prometheus/util/testutil"
 )
 
 // TestWALRepair_ReadingError ensures that a repair is run for an error
