+++ util/testutil/directory.go
@@ -14,8 +14,13 @@
 package testutil
 
 import (
+	"crypto/sha256"
+	"io"
 	"io/ioutil"
 	"os"
+	"path/filepath"
+	"strconv"
+	"testing"
 )
 
 const (
@@ -126,4 +131,52 @@ func NewTemporaryDirectory(name string, t T) (handler TemporaryDirectory) {
 	}
 
 	return
+}
+
+// DirSize returns the size in bytes of all files in a directory.
+func DirSize(t *testing.T, path string) int64 {
+	var size int64
+	err := filepath.Walk(path, func(_ string, info os.FileInfo, err error) error {
+		Ok(t, err)
+		if !info.IsDir() {
+			size += info.Size()
+		}
+		return nil
+	})
+	Ok(t, err)
+	return size
+}
+
+// DirHash returns a hash of all files attribites and their content within a directory.
+func DirHash(t *testing.T, path string) []byte {
+	hash := sha256.New()
+	err := filepath.Walk(path, func(path string, info os.FileInfo, err error) error {
+		Ok(t, err)
+
+		if info.IsDir() {
+			return nil
+		}
+		f, err := os.Open(path)
+		Ok(t, err)
+		defer f.Close()
+
+		_, err = io.Copy(hash, f)
+		Ok(t, err)
+
+		_, err = io.WriteString(hash, strconv.Itoa(int(info.Size())))
+		Ok(t, err)
+
+		_, err = io.WriteString(hash, info.Name())
+		Ok(t, err)
+
+		modTime, err := info.ModTime().GobEncode()
+		Ok(t, err)
+
+		_, err = io.WriteString(hash, string(modTime))
+		Ok(t, err)
+		return nil
+	})
+	Ok(t, err)
+
+	return hash.Sum(nil)
 }
